requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
litellm==1.49.4
python-telegram-bot==21.9